remove redundant glyphs
