+++ packages/jbrowse-plugin-apollo/src/LinearApolloSixFrameDisplay/glyphs/index.ts
@@ -1,3 +1,1 @@
-export * from './BoxGlyph'
 export * from './GeneGlyph'
-export * from './GenericChildGlyph'
